Fix collisions for convex shapes

+++ Source/Core/math/math3d.h
@@ -464,7 +464,7 @@ inline float distance_to_line(vec3 point, vec3 line_a, vec3 line_b)
     float dx = line_a.x - line_b.x;
     float dy = line_a.y - line_b.y;
     float D = dx * (point.y - line_a.y) - dy * (point.x - line_a.x);
-    return D / sqrt(dx * dx + dy * dy);
+    return D / sqrtf(dx * dx + dy * dy);
 }
 
 #endif
